--- a/Cargo.toml
+++ b/Cargo.toml
@@ -9,7 +9,6 @@ keywords = ["topic", "nlp", "phrase"]
 edition = "2018"
 
 [dependencies]
-
 rocket = "0.4"
 serde = "1.0"
 serde_json = "1.0"