@soaxelbrooke 2019-05-12
Thanks for submitting this PR! The sorting and dedup is to reduce non-phrases learned due to repeated messages.  For instance, someone will tweet the same exact message 20 times, which would cause the algorithm to see the amazing co-occurrence of the phrase components to the phrase.

You're correct, this wouldn't change the scoring at all.  I'd love to see a performance comparison of the single core vs per-document parallel strategy you've implemented.  The reason I didn't do per-document parallelism was the general sparsity of n-grams on a per-document basis, which I thought would make it expensive to move across processes compared to the in-place, per-label counting strategy.  Of course, I didn't try it, and I'm not sure what the performance difference is like!