@jjhbw 2019-05-13
Ah, that's a good point; there may be a bit of extra overhead involved as now each document will probably incur some kind of switching overhead. I think this all comes down to what rustc/rayon end up doing with this though, so I have no idea what that would look like.

I've tried quantifying it. Note that the dataset may be a bit too small to properly offset all IO overhead and such. This is on a 4-core `c5.xlarge` on EC2 with:
```
nightly-x86_64-unknown-linux-gnu (default)
rustc 1.36.0-nightly (a784a8022 2019-05-09)
```

# With labels
## Master
```bash
git checkout master
cargo build --release
hyperfine  './target/release/phrase count --mode json assets/reviews.json --textfield body --labelfield category --labelfield sentiment' --warmup 5

Benchmark #1: ./target/release/phrase count --mode json assets/reviews.json --textfield body --labelfield category --labelfield sentiment
  Time (mean ± σ):      1.551 s ±  0.084 s    [User: 4.605 s, System: 0.372 s]
  Range (min … max):    1.473 s …  1.734 s    10 runs
```

## This PR
```bash
git checkout batch-wise-parallelism
cargo build --release
hyperfine './target/release/phrase count --mode json assets/reviews.json --textfield body --labelfield category --labelfield sentiment' --warmup 5

Benchmark #1: ./target/release/phrase count --mode json assets/reviews.json --textfield body --labelfield category --labelfield sentiment
  Time (mean ± σ):      2.132 s ±  0.033 s    [User: 4.277 s, System: 0.232 s]
  Range (min … max):    2.072 s …  2.178 s    10 runs
```

**Conclusion:** as you expected, the document-level parallelism is clearly slower here because it doesn't exploit the highly partitioned nature of the dataset.

# Without labels
## Master
```bash
git checkout master
cargo build --release
hyperfine './target/release/phrase count --mode json assets/reviews.json --textfield body' --warmup 5

Benchmark #1: ./target/release/phrase count --mode json assets/reviews.json --textfield body
  Time (mean ± σ):      1.739 s ±  0.014 s    [User: 1.561 s, System: 0.173 s]
  Range (min … max):    1.715 s …  1.755 s    10 runs
```

## This PR
```bash
git checkout batch-wise-parallelism
cargo build --release
hyperfine './target/release/phrase count --mode json assets/reviews.json --textfield body' --warmup 5

Benchmark #1: ./target/release/phrase count --mode json assets/reviews.json --textfield body
  Time (mean ± σ):      1.311 s ±  0.016 s    [User: 2.156 s, System: 0.300 s]
  Range (min … max):    1.284 s …  1.330 s    10 runs
```
**Conclusion:** the document-level parallelism is clearly faster here, because it can exploit all system threads while the group-level approach cannot. 
I expect this difference to be more pronounced on larger datasets where larger batch sizes make sense. In this example, none of the cores could reach 100% as the batches of work were too small.
Example:

# On my 400MB unlabelled dataset
## Master
```bash
git checkout master
cargo build --release
time ./target/release/phrase count posts.txt

real    3m5.206s
user    3m0.041s
sys     0m4.945s
```

## This PR
```bash
git checkout batch-wise-parallelism
cargo build --release
time ./target/release/phrase count posts.txt

real    1m57.643s
user    3m50.524s
sys     0m9.538s
```