@soaxelbrooke 2019-05-13
Wow, thanks for the thoughtful benchmarking here! I think you're onto something, but something in my gut says that there is a good middle ground that can maintain per-core efficiency, but can still complete single-partition tasks faster.  I'd love it if we could chat more about this here so we can get to the best overall solution, if you can bare with me!

I think the big thing getting in the way of the per-document strategy is the sparsity of natural language - most n-grams will be of count 1, and actually are forced to be that way here to reduce significance being caused by single messages with repeated phrases.  So, what is each thread doing? They're largely just tokenizing (with regex) and aggregating into n-grams (ArrayStrings on the stack), and for a single document, this will be a map from these ArrayStrings to 1 (since there can be no other count for a document due to the "term frequency" strategy used here). If N is our max n-gram size, and M is the mean number of tokens per document, I'd think the things impacting performance most would be:

1. 1 move of whole document (into thread)
2. Regex matching for tokenization
3. M * N copies for copying tokens into n-gram queues
4. M * N copies for recording distinct n-grams
5. 1 allocation and move of the `{ n-gram: count }` map back into the parent thread

2, 3, and 4 happen in both strategies, and 4 happens once per document in the per-document strategy. Though, now that I think about it, perhaps two copies of each token isn't required?  It seems like we should only need to copy a token N times. The n-gram count hashmap has constant time lookup, so counting into a smaller hashmap then merging seems like it would cost twice the total resources of counting into the central count hashmap.  The former, though, allows parallelism in the steps before hand: tokenization and n-gram aggregation.

Perhaps there is a better path in parallelizing everything before the n-gram counting (`f(document) -> distinct_ngrams`), and handling the n-gram counting in the main thread?