@jjhbw 2019-05-14
Interesting analysis. Useful to write it up like this.

However, after having played with this parallel implementation on a larger dataset, i'm convinced that the first thing that should be tackled is the blocking IO. Even with larger batch sizes, full CPU saturation is only very briefly achieved on bigger machines. Come to think of it, this may also improve per-core efficiency and reduce the number of overall syscalls.

I also like think that there are fewer allocs taking place during queue communication than you say, as Rayon / Rustc pass references as pointers across threads. I'm not sure how the stack-allocated data structures come into play here, though. I'm also not familiar with the entire program, i just picked a function that seemed like it would lend itself well to a parallel implementation.